Added get_user_id; fixed add_entry to work for multiple users

+++ tests.py
@@ -6,7 +6,7 @@ import unittest
 """from twilio.rest import Client"""
 
 # Project File Imports
-from database_commands import add_entry, connect, custom_query, today
+from database_commands import add_entry, connect, custom_query, get_user_id, today
 from parsing_logic import parse_expense, shortcuts
 """from send_sms import custom_notif"""
 
@@ -22,25 +22,25 @@ class TestParser(unittest.TestCase):
     date = test_values["date"]
 
     def test_parser_on_cost_only_input(self):
-        self.assertEqual(parse_expense(f"{self.cost}"), (f'{today}', f'{self.cost}', None, None), "Parser is failing on single-value input.")
+        self.assertEqual(parse_expense(f"{self.cost}"), ([f'{today}', f'{self.cost}', None, None]), "Parser is failing on single-value input.")
 
     def test_parser_on_cost_item_input(self):
-        self.assertEqual(parse_expense(f"{self.cost} for {self.item}"), (f'{today}', f'{self.cost}', f'{self.item}', None), "Parser is failing on cost,item two-value input.")
+        self.assertEqual(parse_expense(f"{self.cost} for {self.item}"), ([f'{today}', f'{self.cost}', f'{self.item}', None]), "Parser is failing on cost,item two-value input.")
 
     def test_parser_on_cost_location_input(self):
-        self.assertEqual(parse_expense(f"{self.cost} at {self.location}"), (f'{today}', f'{self.cost}', None, f'{self.location}'), "Parser is failing on cost,location two-value input.")
+        self.assertEqual(parse_expense(f"{self.cost} at {self.location}"), ([f'{today}', f'{self.cost}', None, f'{self.location}']), "Parser is failing on cost,location two-value input.")
 
     def test_parser_on_cost_date_input(self):
-        self.assertEqual(parse_expense(f"{self.cost} on {self.date}"), (f'{self.date}', f'{self.cost}', None, None), "Parser is failing on cost,date two-value input.")
+        self.assertEqual(parse_expense(f"{self.cost} on {self.date}"), ([f'{self.date}', f'{self.cost}', None, None]), "Parser is failing on cost,date two-value input.")
 
     def test_parser_on_cost_item_location_input(self):
-        self.assertEqual(parse_expense(f"{self.cost} for {self.item} at {self.location}"), (f'{today}', f'{self.cost}', f'{self.item}', f'{self.location}'), "Parser is failing on cost,date two-value input.")
+        self.assertEqual(parse_expense(f"{self.cost} for {self.item} at {self.location}"), ([f'{today}', f'{self.cost}', f'{self.item}', f'{self.location}']), "Parser is failing on cost,date two-value input.")
 
     def test_parser_on_cost_item_date_input(self):
-        self.assertEqual(parse_expense(f"{self.cost} for {self.item} on {self.date}"), (f'{self.date}', f'{self.cost}', f'{self.item}', None), "Parser is failing on cost,date two-value input.")
+        self.assertEqual(parse_expense(f"{self.cost} for {self.item} on {self.date}"), ([f'{self.date}', f'{self.cost}', f'{self.item}', None]), "Parser is failing on cost,date two-value input.")
 
     def test_parser_on_cost_item_location_date_input(self):
-        self.assertEqual(parse_expense(f"{self.cost} for {self.item} at {self.location} on {self.date}"), (f'{self.date}', f'{self.cost}', f'{self.item}', f'{self.location}'), "Parser is failing on cost,date two-value input.")
+        self.assertEqual(parse_expense(f"{self.cost} for {self.item} at {self.location} on {self.date}"), ([f'{self.date}', f'{self.cost}', f'{self.item}', f'{self.location}']), "Parser is failing on cost,date two-value input.")
 
 
 class TestShortcuts(unittest.TestCase):
@@ -75,11 +75,21 @@ class TestDatabase(unittest.TestCase):
 
     def test_can_write_to_database(self):
         try:
-            add_entry('9999-01-01', 00, 'test item', 'test location')
+            add_entry(2,'9999-01-01', 00, 'test item', 'test location')
         except Exception:
             self.fail("Can't write to the database.")
         else:
-            custom_query("DELETE FROM expenses WHERE date = '9999-01-01' AND cost = '00' AND item = 'test item' AND location = 'test location'")
+            custom_query("DELETE FROM expenses WHERE user_id = 2 AND date = '9999-01-01' AND cost = '00' AND item = 'test item' AND location = 'test location'")
+
+    def test_add_entry(self):
+        number = "+1234567890"
+        message_body = "00 for test item at test location on 9999-01-01"
+        try:
+            add_entry(*get_user_id(number)+parse_expense(message_body))
+        except Exception:
+            self.fail("add_entry is failing")
+        else:
+            custom_query("DELETE FROM expenses WHERE user_id = 2 AND date = '9999-01-01' AND cost = '00' AND item = 'test item' AND location = 'test location'")
 
 
 """class TestTwilio(unittest.TestCase):
